+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- Feature: Added 'format' flag in Show command to adopt Slack integration
+
 ## 0.6.0
 
 - Refactor: Removed filename from Show command output